requirements.txt
#requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
paho-mqtt
streamlit
docker
kubernetes
dvc
python-dotenv
requests
six
jsonschema==3.2.0
dnspython
openpyxl
boto3
mlflow
confluent-kafka
plotly
watchdog
streamlit-autorefresh
imbalanced-learn
river>=0.21.0
prometheus-client
fastapi
uvicorn